requirements.txt
python-telegram-bot==21.4
python-dateutil==2.9.0.post0
mysql-connector-python==9.0.0
python-dotenv==1.0.1